fix: add explicit searcher sync before close trial [DET-3889] (#1101)
This change fixes a regression introduced within the development for 0.13.0. There was a decision made in #infra-ag to always checkpoint before searcher requested validations. In the event that we're running adaptive and that validation causes a promotion decision, we have to do this for consistency and reproducibility. Changing this order broke existing, implicit behavior in the trial actor. For example, if an asha trial is running and completes a validation that results in its promotion, on old versions of the master (<= 0.12.13) a preclose checkpoint workload would be issued. While the trial is performing this checkpoint workload, the searcher is able to synchronize the operations is wants the trial to do with the trial. On the current version of the master (0.13.0.dev0), since there is a checkpoint before the searcher requested validation, there is no preclose checkpoint to run. Because of this, the searcher is not able to synchronize with the trial before it decides to close because it has nothing left to run. This change makes synchronizing with the searcher explicit rather than implicit, by adding a sendNextWorkload message and using it if the searcher completes a workload that will result in a state change in the searcher.