chore: generate sitemap for docs site.
We now use sphinx-sitemap to generate a sitemap for the docs website. The
sitemap will be published to `docs.determined.ai/<version>/sitemap.xml`, but all
of the URLs it contains will reference `docs.determined.ai/latest/...`. We will
then arrange to publish a `robots.txt` file at the root of `docs.determined.ai`
to point crawlers at the sitemap for the latest version of the docs.