fix: better handle tasks with mixed TLS/non-TLS masters
If the master is listening over both TLS and non-TLS, it will send TLS
connection information to all containers, since it doesn't keep track of
whether each agent connected over TLS or not. But an agent that
connected over non-TLS would then create a task that tried to use TLS on
the non-TLS port, which would fail.

This updates the agent to avoid that by forcibly disabling TLS when the
agent itself didn't use it. It's a bit of a hack, but the mixed
TLS/non-TLS will soon be made impossible on the master side anyway.