fix: support shell with k8s
In k8s, the files we inject into the container are injected via
individual file-level bind mounts, which are effectively read-only in
docker.

As a result, shell-entrypoint.sh was unable to edit the authorized_keys
file in place, so it is now injected as authorized_keys_unmodified and
the modified version is written at runtime.