fix: check that searcher validation metric values are scalars [DET-2134] (#1149)
The validation metric that is specified as the searcher metric is
supposed to take on only scalar numerical values, since we may need to
compare values of that metric for a hyperparameter search. Other parts
of the system (the raw trial details SQL query, at least) also assume
scalar values. However, the property wasn't actually explicitly checked
on input, so it was possible to successfully run an experiment
outputting non-scalar searcher metric values and not have anything go
wrong until a later database query exploded, as long as the searcher
never needed to make any comparisons.

This adds a check in the harness that the searcher metric only takes on
scalar values so that the issue is caught as soon as possible.