fix: return the original tf.keras model when not training. [DET-3718] (#1112)
Previously, we return the wrapped model for both training and
model exporting. When training, the wrapped model is designed
to get the arguments of users' call on compile and fit. However,
when exporting checkpoints, users expect to get the original model.
So fix wrap_model to be no-op when not training.