feat: change scheduler to allow multiple dtrain jobs per node [DET-3819] (#1158)
This change allows the scheduler to fit multiple parallel training jobs to a single node. For example, if you have a 2-agent cluster with 4-slots per agent, 2-slot tasks can be scheduled and all 4-slot tasks will still consume whole agents. There was spike work done to ensure Horovod can handle this in DET-3818.

The code was refactored a bit to handle this - previously findFits was divided into findSingleSlotFit and findMultiSlotFits. This was because in 1-slot tasks are scheduled the same, where n-slot tasks were scheduled differently based on whether they are experiments or commands. This commit changes this to findSingleAgentFit and findMultiAgentFits because now all n-slot tasks, where n fits on a single agent, are scheduled the same and m-slot tasks, where m does not fit on a single agent, are scheduled the same.