feat: override trial network mode to host only when multi-agent (#1167)
This change updates the trial <-> scheduler relationship by adding a field IsMultiAgent to the task spec. This field is used to indicate when a container is started whether its default networking mode should be overridden. This distinction is needed because this should be the case for distributed training but not parallel.