chore: update trialWorkloadSequencer.checkpointModelCompleted to make sure all state changes are snapshotted (#1199)
This change changes how sequencer.snapshotState() is called when the trial workload sequencer completes a checkpoint. Previously, if the checkpoint was a searcher requested checkpoint (e.g. PBT), we did not snapshot the opIdx being incremented, so a failure between that checkpoint and the next searcher op being completed would result in us redoing the checkpoint.