fix: fix a deadlock when gracefully terminate a trial [DET-3879] (#1180)
* chore: clean up trial.go

Move processing messages to different functions and classify them
into scheduler messages, container messages, and etc.

* fix: fix a deadlock in canceling trial

This commit fixes a deadlock in trial actor. It happens
when a trial is terminated gracefully and the trial containers
are still starting. Before all the trial containers are
started, the trial would not send the rendezvous information,
which blocks trial containers sending back completed workload
messages. This issue exists on both the default and k8s resource
providers. In the default resource provider only when the instances
are launched. However, on k8s, containers are started one by one.
There might be a situation that some containers would never start
up. This commit adds a timeout for all graceful trial terminating
and directly handles as killing if not all containers are connected.

The solution is to check if all containers are ready when gracefully
terminating it. If not, rather than waiting for the completed workloads,
kill all the containers. Also, add a timeout to check if the master
has waited too long for all containers to be connected after the first
one i connected.

* chore: remove asserting first completed step in tests

Previously, when canceling experiments, we assert the first
trial has the first completed step, which is an expected
behavior. Now, we change trial to not wait for the first
workload to complete when gracefully terminating it.
Accordingly, we fix the tests to not do such assertions.