fix: use correctly set visible devices for keras/estimators (#1235) [DET-4076]
This change updates estimators, keras and data layer to use hvd.local_rank() to set visible devices when interfacing with the frameworks, not slot ids. Before multiple dtrain jobs could run on a single node, env.slot_ids were correct, but since slot ids are localized (i.e., if outside the container, it is using gpus 0, 2, 4 and 7, inside the slot ids will be 0, 1, 2, 3 but the uuids will match) this resulted in dtrain jobs trying to use slots it couldn't find.