feat: add /api/v1/experiments/id/validation-history [DET-3729] (#1230)
This adds the /api/v1/experiments/id/validation-history endpoint to describe the historical best validations of an experiment. It is a separate endpoint (from the upcoming /api/v1/experiments/id/validations) because expressing it as a combination of sorts and filters isn't realistic (it's a window agg) and it only needs the searcher metric.