feat: expose _TrainContext.from_config() (#1336)
_TrainContext.from_config() creates a training config via the same
mechanism that we use for local test mode.  In effect, this allows users
to instantiate their own Trials and mess with them them in totally
framework-native ways.

An alternate strategy which was considered was to create some sort of
DummyContext, which would have been a totally transparent object that
took a sample of hparams as its input, and which visibly did not
interfere with any objects (i.e. wrap_model() would just return its
input).

The from_config() strategy was chosen because various context functions
(like custom reducers in estimator, or data_layer caching) are useful
to user model code, but would have to be reimplemented with the
DummyContext.  Or, alternately, the user would have to only use a subset
of context methods when they used the DummyContext.  A real context that
should work with any real Trial seems far more useful, at a slight cost
to transparency.