feat: don't log through agent 0 [DET-4180] (#1344)
This change starts Horovod worker processes with their output streams redirected to the output streams of the main process of the container they are in, so that logs do not go through agent zero and are available through standard docker logging.

There were several ways to do the redirection. We could redirect in `worker_process.py`, either with the contextlib or dup'ing the file descriptors or do the redirection in a shell script, or you could do it in a python script (once again by dup'ing file descriptors). A python script was determined to be the simplest, to avoid `contextlib.redirect_stdout` not being respected, packaging woes of a shell script and dup'ing file descriptors in `worker_process.py` being a bit more tricky (in regard to things logging before the swap (imports) resulting in potentially misleading logs).

See https://github.com/moby/moby/issues/19616 for more around what informed the choices made here.