fix: det-deploy deprovisions GCP agents despite long master names [DET-4271] (#1391)
The provisioner default label_value for GCP is truncated to 30
characters. This was problematic for deprovisioning dynamic agents since
det-deploy looks up instances using the expected default label key /
value and failed to truncate in the same way. This truncation will be
relaxed in the future, but for now, det-deploy works around this by
explicitly setting label key/value in the cluster configuration.