feat: add opentracing to actor system [DET-4212] (#1327)
This change adds distributed tracing support to the actor system, to better visualize and understand the actor system and how parts of the system interact. This is preliminary support. It does not implement a websocket carrier to carrier span contexts over websockets, to and from the agents and harness. However, the trace is complex as is and that's probably not necessary.