fix: improve performance of agents endpoint for k8s [DET-4073] (#1450)
Previously when we accessed the agents endpoint (WebUI, det slot list, det agent list)
while running on k8s, we would query the k8s API server for the current list of nodes
which was a blocking call. We now keep a running list of current nodes, so there is no
need to make blocking calls to the k8 API server while building the response for the
agents endpoint.