refactor: clean up code for master-sent messages
A bunch of code relating to the `TaskSpec` struct was left over from
when that type was serialized and sent to the old Python agent; that
code is now removed, since `TaskSpec` is no longer sent directly but
instead translated into a `container.Spec` within the master.

This also deletes the unused `KillContainer` and moves the message for
running a workload, which was stuck in among all the container start
messages, to a more appropriate place in the trial code.