chore: add new fields to trial logs [DET-4176] (#1373)
This change adds new columns that will be used for filtering trial logs to the trial logs table in the database and migrations to create them. Additionally, a new message column is added for new log messages written by Fluentd primarily so they can be reconstructed to match previous trial logs.