feat: make slot type configurable. [DET-4308] (#1484)
Previously, we only support using autoconfiguring the slot type to be GPU and if there are no GPUs fall back to CPU. This applies to most use cases, including GPU-only cluster and testing infrastructure (typically users don't use purely CPU). However, we want to be able to support zero-slot resource pools for zero-slot tasks, such as Tensorboard, GC tasks, CPU-only notebooks, and etc,. And we don't want zero-slot resource pools to show any slots on the slot utilization.

Now, in order to support a zero-slot resource pool when we support multiple resource pools, we need to make slot type configurable for agents as opposed to always autoconfiguring the slot type. For dynamic agents, we always use GPUs as slot types so resource pools that use CPU agents will be considered to only run zero-slot tasks.