feat: use Fluent Bit for trial logs [DET-4178] (#1462)
This PR moves trial logs to being sent through a Fluent Bit container
running alongside each agent in the cluster. For the moment, doing so
mostly just adds complexity, but the idea is to later add support for
Elastic as a separate log backend, which will be made easier by the use
of Fluent Bit.

The agent internally manages the lifecycle of the Fluent Bit container
in order to keep administration simple and ensure that the daemon is
started with the necessary configuration. The agent automatically
configures Fluent Bit to connect back to the master and fill in the new
fields being used for filtering. To simplify the required networking,
the logs that the agent itself inserts are sent to the master rather
than through Fluent Bit.

The master gains a new endpoint to accept POSTed trial logs, which it
sends over to the trial logger actor. Only trial logs are sent through
this new path; in the long term, that should also be unified with
commands and other tasks, but those aren't the same currently anyway, so
that's being left out of this project.