fix: allow clients to override the expected master cert address [DET-4547] (#1588)
To handle cases where the address the agent is using to connect to the
master is not one of the valid addresses for the master certificate
(e.g., public vs. private addresses of a cloud instance), this adds the
ability to the various parts of the stack to option to expect a
different address in the certificate from the address that is actually
being used to connect.