test: fix CircleCI timing-based splitting for E2E tests (#1622)
pytest was producing JUnit XML files in a slightly wrong format, causing
CircleCI to fail to find timing information for E2E tests and preventing
it from doing the automatic timing-based test splitting we wanted it to
do. This fixes the format so that the splitting works, reducing the time
for running E2E CPU tests from 20 minutes to 12, roughly. That doesn't
totally translate into a reduction in time for the overall E2E workflow,
since the affected job is no longer the longest one, but total time is
down by 5 or so minutes.