chore: fix fluent lua filter (#1642)
This change fixes the behavior of the lua filter run by fluent bit when it receives a nil log entry. In the event a log like [rank=0] was received prior to this change, the rank parser would parse out rank as 0 and log as nil according to the regex ^\[rank=(?<rank_id>([0-9]+))\] (?<log>.*). This leads to the lua snippet in a later filter to try to concatenate nil .. '\n' which results in an error. It seems even though the filter errors, fluentbit considers the filter optional and attempts to ship the log without running it. Since the master deserializes batches of logs together and the lua filter converts rank id to an int, this caused the master to receive malformed JSON and throw out the entire batch resulting in more logs being lost than would be anticipated by this bug.