feat: support max zero slot containers for resource pools [DET-4309,DET-4340] (#1507)
This PR fixes scheduling CPU tasks and scaling for CPU resource pools. It fixes the fitting methods because the current fitting methods are not friendly to zero-slot tasks. It fixes calculating the number of agents because previously zero-slot tasks are not considered into calculating the agent number for scaling.

It also adds a field max_zero_slot_containers_per_agent in each resource pool configuration to control the scheduling and scaling behavior on CPU agents.

In the meantime, related tests are fixed and more test cases are added for zero-slot tasks and cpu resource pools.