feat: support elasticsearch as a trial logging backend [DET-4179] (#1542)
The change allows trial containers to use elasticsearch as a logging backend. It also adds a mechanism for passing extra configurations to agents on their first connection to the master. This is necessary to securely pass credentials, such as the username and password for elasticsearch, to the agent within the constraints of some secure deployments.

Commentary included with the PR is included below.

How did we pick the elastic index?
This just using the standard log stash style indexing. The index logs are shipped to an index like triallogs_11-02-2020 that rolls over each day. When the first log is shipped, the index is autocreated. Users should configure elastic to use a reasonable number of shards per auto created index by using index templates (https://www.elastic.co/guide/en/elasticsearch/reference/current/index-templates.html for indices that match our logstash format index names).

How are agents getting the elastic configuration?
As is, the agent starts, connects to the master and receives the configuration back in a new message. This comes as the alternative to to the "obvious", which is putting it in the agent/provisioner config and passing it through in the startup script for the provisioner case and user specified for the static case. This issue with this is templating the variables into the startup script will expose then in gcp/aws dashboards. The alternative workaround to that would be to do something like source the startup script and pass any DET_... variables through to docker (so users can set them securely in their startup script), but this felt even worse, mainly because we'd need a lot of documentation to guide users through this and it would be totally non-obvious to them unless they could find the documentation.