feat: priority scheduler with preemption [DET-4512] (#1634)
Adding priority scheduler with pre-emption. This scheduler is not yet configurable without master
code changes as we are not yet ready to release it.

The priority scheduler will never scheduler lower priority (non-zero slot) task while a higher priority
task is pending. It is able to run with and w/o pre-emption. When pre-emption is enabled, lower
priority tasks will be terminated in order to schedule higher priority tasks. By default pre-emption
is disabled.