test: integration test trial logs API (#1456)
This change adds an integration tests target to the master and integrations tests for the trial logs endpoint in the new api.

Commentary from the PR is included below:

There are a ton of ways to do just this so I'll try to address some probably-incoming questions upfront.

How did you decide to do it this way?
I've structured tests like this in previous projects and it led to good and effective coverage. Also, this pattern is sort of similar to what k8s, containerd, docker and a couple of other projects I perused do. The can also be moved in tree and use the //build +integration tags and work essentially the same way (apparently it seems this is the pattern soundcloud uses http://peter.bourgon.org/go-in-production/#testing-and-validation, and I like it). I'm up for debate on this.

Why start an entire master for each test?
If multiple tests are similar enough to share a master, they can be expressed as subtests and invoked with t.Run(). This allows us to start masters with different configurations for each test to test specific integrations. For example, with them written this way, when I write tests to cover filtering with elastic backed trial logs, I'll be able to run a test suite for these with a properly configured master.

Why start the master in a goroutine instead of a subprocess via the built binary?
This was mainly to introspect the state of the master if needed, but with how complex closing the master got (along the way I found a bug in cmux) I'm not 100% sure it was worth it anymore.

Why are they called integration tests, they look more like e2e tests of test master?
Specifically these tests were added to test the integration with the database and the sort-of complex code to construct filters. They start and close an entire master for each test, but I plan on using the down the line (probably in my next tech debt week to add real experiment/trial tests) so I wanted to go ahead and add it in instead of some in between (this could've also been tested with a master with just a DB and actor system, but then I need to figure out how to run the migrations before the tests start too, if I'm not calling Run()).

How long do they take?
Startup for a test is 500-1000ms and once started the subtests run in <50ms usually. The CI target seems to take 2-3m but most of the overhead is installing some pre-reqs and when go test compiles the tests. I think this can stay significantly shorter than e2e with way more tests.