fix: allow dots in map config keys (#1665)
This change allow keys in map sections of the `master.yaml` to be specified with a `.` in them. This was previously not possible because, while we marshal the configuration successfully into a `map[string]interface{}`, when we merge this configuration into viper, viper loses its ability to disambiguate keys with `.`'s and nested objects because it uses `.` as the object delimiter by default (explained in more detail in comment in code).