fix: let shells work through an HTTP proxy/load balancer
We were using HTTP CONNECT to proxy SSH connections to shells through
the master, but that isn't widely supported by HTTP proxies/load
balancers (in particular, AWS Elastic Load Balancing is documented not
to support it and GCP's load balancing was seen to fail; nginx, as
another example, doesn't support it without an external module).
Instead, we now proxy TCP connections over WebSockets, which have better
support and which we already use in such situations. There's a little
overhead, but it shouldn't be too much: WebSockets are capable of
transmitting binary data, so there's just a bit of extra framing around
the raw data.