fix: limit concurrent restores to avoid resource exhaustion [DET-4556] (#1666)
Limit the number of concurrent restores at any time within the system to maxConcurrentRestores.
This avoids an issue where the connection pool can be exhausted and the system deadlock. For
example if there are N connections in the pool and N experiments restoring, since a connection
is pinned per restore streaming events, if during any restore blocks for a connection before
completing, we are deadlocked.