feat: expose some Fluent Bit logs in the agent (#1680)
If the agent's subordinate Fluent Bit container has an issue, it can be
awkward to debug, since its logs are not exposed in our product and can
even vanish entirely if the container exits. Here, we make the agent (1)
print out any error or warning lines it finds in Fluent Bit's output and
(2) dump all recent logs if the container exits.