feat: read logs from elasticsearch [DET-4621] (#1637)
This PR adds an support for reading logs from Elasticsearch when it is configured as the log backend for the cluster. 

The API is identical to when backed by Postgres, however in several cases this API is not well suited to Elasticsearch. For example, when the WebUI makes queries to the API like `/api/v1/trials/:trial_id/logs?limit=1000&offset=-1000&follow=true`, they will get converted to a large from+size that would exceed index.max_result_window or worse cause the cluster to OOM. This will be addressed by a follow up to support ORDER BY in the API and return better errors in these cases.