chore: log through Fluent Bit on Kubernetes [DET-4622] (#1712)
This makes logging on Kubernetes go through Fluent Bit by creating a
sidecar container running Fluent Bit for each trial pod that shares a
volume with the main container. Logs go to files in the volume and are
picked up by Fluent Bit's file tailing input. In order to limit disk
usage, logs are piped, via shell shenanigans in order to get stdout and
stderr treated separately, through a tool called multilog, which handles
rotating output files based on size. (As opposed to using another
container running logrotate or similar, doing it directly on the output
side avoids messing around with truncating files in place and the like.)

The logs volume is configured to be backed by memory, so logs don't have
to hit a disk just to get from one container to another.