test: fix race condition in test-intg-agent (#1741)
This change fixes two race conditions with the agent integration tests. One involved us not waiting for the container to exit before checking the logs. The other involved logs showing up out of order and from different fluent flushes. This is fixed by awaiting the logging container exit and just waiting for all the logs and resorting them when checking if they were received.