chore: hack tensorboard support to include custom metrics (#1750)
Our Tensorboard support is still based on the totally outdated Callback
class (BatchMetricWriter is the only remaining implementation of
Callback).

The bug was that PyTorch custom reducers for training metrics did not
show up in Tensorboard at all.  This is because they are only set in
avg_metrics, and the Callback interface never used to pass avg_metrics.

Since there is only one implementation of Callback, we can just change
pass batch_metrics and avg_metrics to allow BatchMetricWriter to detect
these custom reduced metrics.  It would be preferable to break the
Tensorboard logic out of the WorkloadManager, but this hack is easy
enough to land in time for the feature release of PyTorch custom
reducers.