fix: workaround boto3+minio bug (#1770)
I am not sure when this broke (or how it ever worked actually), but this is a workaround for boto/botocore#1328 (I think) that causes our system to fail when using MinIO as an s3 provider.

The workaround is to detect when MinIO is in use (via HTTP headers) and simply avoid uploading empty directories. The only reason that the S3 StorageManager uploaded empty directories anyway was to imitate the behavior of the S3 console, which is not a very big deal.