feat: json-schema for experiment config validation (#1715)
Json-schema defines validation logic for arbitrary schemas with good
cross-language support.  This allows us to migrate our experiment config
validation logic (and defaulting logic) out of Go and into both Python
and JavaScript (this PR only includes the Go and Python validation).

Json-schema is not natively capable of all of the validation logic that
we need, but it is still highly valuable for our usecase because custom
extensions are a first class feature in most json-schema libraries.

We have the following custom extensions:
  - checks: Custom messages for arbitrary validation logic
  - compareProperties: Support customizable value comparisons
  - conditional: Enforce a schema whenever some condition is met.
  - disallowProperties: Custom messages when disallowing properties
  - eventuallyRequired: Support two-step validation
  - union: Excellent error messages when validating union types

This PR's effect on the current system is purely cosmetic.  The new
json-schema validators are used purely for improving the invalid config
error messages, but all validation and defaulting logic otherwise
remains the same.

The goal of this PR is to have identical logic as before but in a
cross-language format.  The only exceptions are where the old behavior
is completely wrong (additional fields are silently allowed in the
ResourcesConfig).  A non-goal is to modify the experiment config schema
from its current state.  Eventually, supporting automatic shimming of
experiment configs between explicit versions will give us the
flexibility to modify our existing experiment config schema as we wish.