chore: return consistent log IDs [DET-4789] (#1775)
This change alters the trial logs API to return globally consistent IDs, instead of IDs consistent within the context of the series of requests the WebUI used to make to this API. The IDs are returned directly from the log backend, which means they must be returned as a protobuf one of since postgres uses int IDs where elasticsearch uses strings. This looks like a breaking change but, fortunately, it only breaks protobuf not the JSON provided by grpc gateway.