fix: always encode trial ID as string [DET-4789] (#1800)
This changes the type of api.v1.TrialLogsResponse.id to be a string as opposed to an int. Previously, this field was a oneof with int ids in the case of Postgres-backed logs and string in the case of Elasticsearch, but something broke down in the encoding and the int64 oneof ID's returned from the JSON API provided by grpc-gateway were erroneously returned as string. In place of debugging where that break down happens, just changing the encoding is simpler and a better API.