chore: enable some more Go linters (#1839)
This change enables the following golangci-lint linters:

- exportloopref: prevent the bug that was fixed by #1834
- gochecknoinits: reduce the side effects of imports
- godot: help comment formatting conform with standard Go conventions

The calls to `rand.Seed` and `petname.NonDeterministicMode` are
technically unnecessary, since `github.com/labstack/gommon` internally
seeds the global random `math/rand` generator on import and they all go
to the same place. But that should be treated as an implementation
detail and the multiple seeding is harmless, though pointless.