fix: another place scheduler config for resource pool not being inherited (#1854)
Fix bug where scheduler settings for a resource pool are filled in by default. Correct behavior should allows the scheduler config to be empty, which indicates that the resource pool should inherit the settings from the ResourceManager's settings.

This PR is a follow on to #1847, which missed one location where the scheduler is being configured with defaults where it should actually be left unconfigured. This PR fixes that mistake.