feat: swap master restart to be snapshot based (#1745) [DET-816]
This change removes the event log and changes master restart to instead be snapshot based.
This involves but is not limited to serializing the state of experiments, searchers, trial workload
sequencers and trials, saving these states in a debuggable format, saving these states from
distributed actors in a consistent manner, restoring from these states, and other bits, like garbage
collecting and logging. An explanation of the new restoration system is included below and also
in the code as commentary.

Restore works by restoring from distributed consistent snapshots taken through the course
of an experiment. Snapshots within the system flow from the bottom up, starting with the
trial workload sequencer, to the trial, and finally to the experiment. Any event that the
trial or trial workload sequencer processes that would trigger a change to the state of the
experiment is:

    Propagated atomically, within a single message, to ensure the experiment handles it all
    or nothing
    With a snapshot affixed to it, to mark that it should trigger a snapshot

Upon receipt, the experiment handles the event entirely, snapshots its state and saves it
along with all the snapshots it has received, atomically.

This is technically equivalent to the Chandy-Lamport distributed snapshotting algorithm where
the trial workload sequencer initiates the snapshots and propagates them along with trialCreated
and trialCompletedWorkload messages, with the snapshot itself acting as the marker. Each actor we
care about and know has updated state in the system snapshots it state when it receives this
marker and propagates it through the system. Even though all trials are part of each distributed
snapshot, we do not propagate snapshots to trials that didn't initiate the snapshot; that a trial
sends a snapshot with each state change implies if it did not send the snapshot, it does not have
a state change [1]. We can also be sure there are no pre-snapshot messages floating in the ether,
since any message we care about snapshotting is itself a snapshot.

To restore properly, any message the experiment would send as a result of snapshot-able event
must be replayed on restore. In our case, these are the new operations created by the searcher
as a result of a trialCreated or trialCompletedWorkload message. We restore trials and experiment
directly from their snapshots and tell the trial all its operations on creation.

[1] This is fine, but also we have to make sure that nothing else alters trial and experiment
state that doesn't flow from trial to experiment. For example, right now experiments update the
trials with what their state is. If we wanted to snapshot this, we would need to move to a full
Chandy-Lamport implementation since as is we assume the trial has no state updates between
snapshots and this breaks that invariant.