fix: CI failures caused by resource pool merge (#1864)
Update the CircleCI config to use the newly renamed arguments to
det-deploy.

Change test that was designed around the assumption that the experiment
and the TensorBoard will run on the same instance from E2E GPU (where
the experiment will run in the GPU resource pool and the TensorBoard
will run in the CPU resource pool) to E2E CPU.

Fix bug in the code where TensorBoards were incorrectly being routed to
the default GPU resource pool when TensorBoards are always zero-slot
tasks. This was caused by all commands defaulting to slots=1, the
default resource pool being chosen based on whether slots=0, and the
TensorBoard manager not setting slots=0 until AFTER the default resource
pool was set.