chore: let CLI verify the master using combined system/custom certs (#1859) [DET-4666]
Previously, when the CLI found a stored custom TLS certificate for the
master, it would verify all requests to the master using exactly that
certificate. That worked well enough for a single master, but led to a
thorny development experience when switching between different masters:
with a custom certificate around, the CLI could no longer connect to a
master using a certificate signed by a well-known CA without an
additional prompt (and often not at all, since the verification
available in the CLI needs a full chain to the root and root certs are
usually left out of chains in practice).

This change avoids that issue by instead constructing and using a
temporary bundle that contains both the custom bundle and all well-known
CA certs (as defined by Certifi, which Requests defaults to
normally). For now, we refrain from handling multiple custom
certificates.

This may be a bit dirty. An alternative technique would be to dive into
Requests/urllib3 and do some jiggery-pokery with custom transport
adapters and SSL contexts or something, but I looked into that and
decided it would be much too hairy to be worth it.