feat: support more fields to sortBy in /api/v1/experiments/trials [DET-4219, DET-4920] (#1899)
This change adds support to sort by validation end time, validation metric, duration and total batches processed to `/api/v1/experiments/$1/trials` and fixes a bug with the endpoint that caused none of the sort by's to work. It adds the SQL function public.page_info that replicates the behavior of api.Pagination() in Go code and fixes a regression in `/api/v1/experiments` where using offset and limit as provided didn't work as expected.