refactor: reduce code duplication across task types (#1749)
For a long time, we've had largely separate code paths with a bunch of
duplication to handle each of the three top-level task types (trials,
checkpoint GC, and commands). That started with the agent code and
transferred over into Kubernetes as well. This change unifies the code
paths to operate on a generic task spec type that has methods providing
the parameters that differ between tasks; everything only operates on
that type generically.

`TaskSpec`, rather than containing a pseudo-union (three pointers,
exactly one of which is expected to be non-nil), now holds one instance
of a new interface that is implemented by a concrete struct type for
each of the task types.