refactor: replace step ID with total batches in the DB (#1890)
Replace the use of step id with the total batch in the DB. Any checkpoint, training metric, validation can be located with a trial id and a total batch number. They used to depend on the concept of steps. In order to hand over the control on checkpointing and validating to the user code, all the internal use of steps should be abandoned. In the meantime, remove the constraint that all checkpoints and validations need to be associated with a step.