feat: allow retaining CloudWatch log groups of det-deploy clusters (#1948)
`det-deploy aws up` now supports two new options: `--retain-log-group`
(only on the simple template) and `--log-group-prefix`. The former sets
the CloudWatch log group that is created for the stack to not be deleted
when the rest of the stack is, which may be useful in some
situations. The latter sets the prefix that is applied to the stack name
to generate the log group name (default "determined", leading to the
same behavior as before).

Because CloudFormation's templating is terrible, the implementation
looks a bit awkward. The `DeletionPolicy` field is required to be a
literal, so it is not possible to use an expression to set it
conditionally. Instead, we have to define two separate resources, one
retained and one not, each of which is created conditionally. As far as
I can tell, those conditions can't be written inline but instead have to
be defined in a separate section and referred to by name, which also
means we need one for each case. Having two separate resources in turn
means that we can't refer to the log group by its logical ID later on
when we need the log group name; instead, we have to insert the
expression used to compute the log group name everywhere we need
it (since it also seems to be impossible to define derived variables in
the template without doing something crazy like going through AWS
Lambda or storing config snippets in S3).