ci: set a CI-specific prefix for and retain CloudWatch log groups
In order to avoid the (somehow) error-prone step of exporting CI logs
from CloudWatch to S3, we now set logs from CI-spawned det-deploy AWS
test clusters to be retained after the CloudFormation stacks are
deleted. Those log groups also now have a new "determined-ci" prefix to
separate them from other log groups, now that they're going to stick
around.

This also includes a fix for an issue in updating a stack after the
recent det-deploy change introducing log retention: the logical name of
a deleted log group is now set to the same as it was before, which
should avoid a conflict in CloudFormation.