chore: support task tokens in harness for authentication [DET-4897] (#1894)
All the harness internal APIs will access the restful APIs with authentication. It uses a task token that is passed into the container. All task tokens' lifecycles are tied with the tasks' lifecycles.

Specifically, all the requests can have an HTTP header key-value pair with the key Grpc-Metadata-x-task-token. The gRPC server on the master will truncate the prefix Grpc- in the HTTP header and look it up in the task session list that exists in the database.