fix: use UTC for all logs (#1985)
This change amends all the logging-related code where timestamps with timezones were used instead of UTC time. It fixes a bug where the master and agents wrote logs with a timezone into a Postgres timestamp (without timezone, which apparently just doesn't work at all, the UTC+1 logs appear as if they are UTC+0 but an hour in the future; the timezone info is just truncated not converted to UTC), which caused the few UTC+N logs from the master and agent to get returned over and over again.