fix: fix query for GET experiments [DET-5039] (#2012)
This change updates the GET experiments query to handle null jsonb (not just a NULL column) when filtering by labels. The labels column initially is returned as NULL if no labels key exists in the config jsonb. The expression works fine in this case, but if labels are added to an experiment and the removed it will get rewritten as null instead of removing the column which causes jsonb_array_elements_text to barf on the call with the arg null jsonb. This PR just checks for null jsonb and converts it to NULL.