feat: add aggregated resource allocation endpoint and job
We now support returning cluster allocation metrics aggregated by
day. It seems difficult, after some experimentation, to do aggregation
on the fly sufficiently quickly in the face of very large histories, so
the aggregation is done at the end of each day and the endpoint then
does a fast query over those results. There's a new aggregation actor
that stays alive in the background and updates the aggregated table at
the end of each day or on master startup.